+++ scripts/osx/dist-win.sh
@@ -1,18 +1,19 @@
 #!/bin/bash
 STARTING_DIR=$(pwd)
 TARGET_DIR=~/Projects/yhat/rodeo
-NODE_VERSION=6.4
 
 cd $TARGET_DIR
 
+rm -rf dist
+
 #dependencies
 ./scripts/osx/install-deps.sh
-./scripts/osx/install-dist-deps.sh
+./scripts/osx/install-dist-win-deps.sh
 
 #remember nvm
 echo '#guarantee nvm'
 source $(brew --prefix nvm)/nvm.sh
-nvm use $NODE_VERSION
+nvm use
 
 #build distributable
 ./node_modules/.bin/gulp dist:win
