+++ src/browser/jsx/containers/history-viewer/history-viewer.reducer.js
@@ -5,6 +5,9 @@ import immutableUtil from '../../services/immutable-util';
 import errorsService from '../../services/errors';
 import cid from '../../services/cid';
 import textUtil from '../../services/text-util';
+import reduxUtil from '../../services/redux-util';
+
+const prefix = reduxUtil.fromFilenameToPrefix(__filename);
 
 /**
  * If any of the history blocks are jupyterResponse types, then they might need to be updated with new content
@@ -128,10 +131,47 @@ function databaseConnectionQueried(state, action) {
   return state;
 }
 
-export default mapReducers({
-  HISTORY_VIEWER_CONTRACT: contract,
-  HISTORY_VIEWER_EXPAND: expand,
+/**
+ * @param {Array} state
+ * @param {object} action
+ * @returns {Array}
+ */
+function blockAdded(state, action) {
+  const blockIndex = _.findIndex(state.blocks, {id: action.block.id});
+
+  if (blockIndex === -1) {
+    state = state.set('blocks', state.blocks.concat([action.block]));
+  }
+
+  return state;
+}
+
+/**
+ * @param {Array} state
+ * @param {object} action
+ * @returns {Array}
+ */
+function blockRemoved(state, action) {
+  const blockIndex = _.findIndex(state.blocks, {id: action.blockId});
+
+  if (blockIndex > -1) {
+    let blocks = state.blocks.asMutable();
+
+    blocks.splice(blockIndex, 1);
+
+    state = state.set('blocks', blocks);
+  }
+
+  return state;
+}
+
+export default mapReducers(_.assign(reduxUtil.addPrefixToKeys(prefix, {
+  CONTRACT: contract,
+  EXPAND: expand,
+  BLOCK_ADDED: blockAdded,
+  BLOCK_REMOVED: blockRemoved
+}), {
   JUPYTER_RESPONSE: jupyterResponseDetected,
   DATABASE_CONNECTION_QUERYING: databaseConnectionQuerying,
   DATABASE_CONNECTION_QUERIED: databaseConnectionQueried
-}, {});
+}), {});
